requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
pydantic>=2.6
python-dotenv>=1.0
pandas>=2.0
openpyxl>=3.1
langchain>=0.2.16
langchain-community>=0.2.16
langchain-text-splitters>=0.2.4
langchain-chroma>=0.1.4
chromadb>=0.5.5
langchain-google-genai>=1.0.8
pypdf>=4.2
docx2txt>=0.8
